backend/requirements.txt
# Core Django dependencies
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8

# WebSocket support
channels==4.0.0
channels-redis==4.1.0

# HTTP client for API calls and testing
requests==2.31.0

# ASGI server for production
daphne==4.0.0

# Environment configuration
python-dotenv==1.0.0

# File handling and uploads (optional but commonly used)
Pillow==10.0.1

# Storage backends
django-storages==1.14.2
boto3==1.34.0  # For S3 storage

# Database (if upgrading from SQLite)
# psycopg2-binary==2.9.7  # For PostgreSQL
# mysqlclient==2.2.0      # For MySQL

# Development and testing
coverage==7.3.2
factory-boy==3.3.0

# Production server (choose one)
# gunicorn==21.2.0        # For WSGI apps
# uvicorn==0.23.2         # Alternative ASGI server

# Caching (for production)
# redis==5.0.1

# API documentation (optional)
# drf-spectacular==0.26.5
